requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.8
Markdown==3.0.1
mysqlclient==1.3.14
nameparser==1.0.2
Pillow==5.4.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
